+++ Lib/EEPROM.h
@@ -1,7 +1,7 @@
 #ifndef EEPROM_H
 #define EEPROM_H
 
-#include "Board.h"
+#include "STM32X.h"
 
 /*
  * EXAMPLE BOARD DEFINITION
@@ -19,8 +19,8 @@
  * PUBLIC FUNCTIONS
  */
 
-void EEPROM_Write(uint32_t offset, const void * data, uint16_t size);
-void EEPROM_Read(uint32_t offset, void * data, uint16_t size);
+void EEPROM_Write(uint32_t offset, const void * data, uint32_t size);
+void EEPROM_Read(uint32_t offset, void * data, uint32_t size);
 
 /*
  * EXTERN DECLARATIONS
